Extract and test makeCosmoshubPath
