+++ packages/launchpad/src/index.ts
@@ -124,6 +124,7 @@ export {
   MsgWithdrawDelegatorReward,
   MsgWithdrawValidatorCommission,
 } from "./msgs";
+export { makeCosmoshubPath } from "./paths";
 export {
   decodeAminoPubkey,
   decodeBech32Pubkey,
@@ -137,6 +138,6 @@ export { AccountData, Algo, AminoSignResponse, OfflineSigner } from "./signer";
 export { CosmosFeeTable, SigningCosmosClient } from "./signingcosmosclient";
 export { isStdTx, isWrappedStdTx, makeStdTx, CosmosSdkTx, StdTx, WrappedStdTx, WrappedTx } from "./tx";
 export { pubkeyType, PubKey, StdFee, StdSignature } from "./types";
-export { makeCosmoshubPath, executeKdf, KdfConfiguration } from "./wallet";
+export { executeKdf, KdfConfiguration } from "./wallet";
 export { extractKdfConfiguration, Secp256k1HdWallet } from "./secp256k1hdwallet";
 export { Secp256k1Wallet } from "./secp256k1wallet";
