+++ packages/launchpad/src/wallet.ts
@@ -1,27 +1,11 @@
 import {
   Argon2id,
-  HdPath,
   isArgon2idOptions,
   Random,
-  Slip10RawIndex,
   xchacha20NonceLength,
   Xchacha20poly1305Ietf,
 } from "@cosmjs/crypto";
 import { toAscii } from "@cosmjs/encoding";
-
-/**
- * The Cosmoshub derivation path in the form `m/44'/118'/0'/0/a`
- * with 0-based account index `a`.
- */
-export function makeCosmoshubPath(a: number): HdPath {
-  return [
-    Slip10RawIndex.hardened(44),
-    Slip10RawIndex.hardened(118),
-    Slip10RawIndex.hardened(0),
-    Slip10RawIndex.normal(0),
-    Slip10RawIndex.normal(a),
-  ];
-}
 
 /**
  * A fixed salt is chosen to archive a deterministic password to key derivation.
